backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
pydantic
python-dotenv
requests
pandas
numpy
scikit-learn
statsmodels
yfinance
alpha_vantage
newsapi-python
tweepy
python-multipart
python-jose[cryptography]
bcrypt
psycopg2-binary
pandas-ta
xgboost
tensorflow
